backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
billiard==4.2.4
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.109.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
kombu==5.6.1
motor==3.3.2
ndg-httpsclient==0.5.1
packaging==25.0
passlib==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==12.0.0
prompt_toolkit==3.0.52
pusher==3.3.3
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
PyNaCl==1.6.1
pyOpenSSL==25.3.0
PyPDF2==3.0.1
pypdfium2==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.27.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==12.0
